Project structure changed 1. CIMController divided to 2 classes CIMService and CIMController(will refactored later) 2. Refactored dependencies 3. ConnectionView and ConnectionViewController extracted to 2 files
