+++ cim_client/lib/views/get_controller.dart
@@ -1,19 +1,12 @@
 import 'package:cim_client/CIMDataProvider.dart';
 import 'package:cim_client/CIMUser.dart';
-import 'package:cim_client/view/cim_connection.dart';
-import 'package:cim_client/view/get_view_model.dart';
+import 'package:cim_client/views/cim_connection.dart';
+import 'package:cim_client/views/get_view_model.dart';
 import 'package:get/get.dart' hide Trans;
 import 'package:cim_client/globals.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
-
-
-enum CIMViews{
-  authorisation_view,
-  connection_view,
-  main_view
-}
 
 enum MainMenuItems{
   item_patients,
@@ -24,12 +17,12 @@ enum MainMenuItems{
 
 class Controller extends GetxController{
 //class Controller extends GetxService{
-  CIMConnection connection;
+//  CIMConnection connection;
   CIMDataProvider dataProvider;
   Rx<MainMenuItems> selectedItem;
-  Rx<CIMViews> currentView;
+
   AuthorisationViewModel authViewModel;
-  ConnectionViewModel connectionViewModel;
+//  ConnectionViewModel connectionViewModel;
   PatientsScreenModel patientsScreenModel;
   RxBool authorised;
   bool isAuthorised() => authorised.value;
@@ -37,10 +30,10 @@ class Controller extends GetxController{
     //TODO: Implement authorisation procedure
     authorised.value = true;
   }
-  void onConnectionChanged(){
+/*  void onConnectionChanged(){
     currentView.value = CIMViews.authorisation_view;
     //TODO: Implement connection
-  }
+  }*/
 
   void onSelectMainMenuItem(MainMenuItems item){
     if(item == selectedItem.value)
@@ -48,41 +41,15 @@ class Controller extends GetxController{
     selectedItem.value = item;
   }
 
-  void checkConnection(){
-    connection.address = connectionViewModel.address;
-    connection.port = connectionViewModel.port;
-    connectionViewModel.connectionState = ConnectionStates.checking;
-    connection.connect().then((value){
-      if(value == 0){
-        connectionViewModel.connectionState = ConnectionStates.connected;
-      }
-      else{
-        connectionViewModel.connectionState = ConnectionStates.disconnected;
-        String message = mapError[value].tr();
-        Get.defaultDialog(
-            title: "error".tr(),
-            middleText: message,
-            confirm: RaisedButton(
-              child: Text("OK".tr()),
-              onPressed: ()=>Get.back(),
-            ),
-        );
-      }
-    });
-  }
+
 
   @override
   void onInit() {
     super.onInit();
     dataProvider = CIMDataProvider();
     selectedItem = Rx(MainMenuItems.item_patients);
-    currentView = Rx(CIMViews.authorisation_view);
     authViewModel = AuthorisationViewModel();
 //    connection = Get.put(CIMConnection());
-    connection = CIMConnection();
-    connectionViewModel = ConnectionViewModel();
-    connectionViewModel.address = connection.address;
-    connectionViewModel.port = connection.port;
     patientsScreenModel = PatientsScreenModel();
     authorised = false.obs;
   }
